+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/PolarisConfigBootstrapAutoConfiguration.java
@@ -20,7 +20,6 @@ package com.tencent.cloud.polaris.config;
 
 import com.tencent.cloud.polaris.config.adapter.AffectedConfigurationPropertiesRebinder;
 import com.tencent.cloud.polaris.config.adapter.PolarisConfigFileLocator;
-import com.tencent.cloud.polaris.config.adapter.PolarisPropertySourceManager;
 import com.tencent.cloud.polaris.config.condition.ConditionalOnReflectRefreshType;
 import com.tencent.cloud.polaris.config.config.PolarisConfigProperties;
 import com.tencent.cloud.polaris.config.config.PolarisCryptoConfigProperties;
@@ -60,15 +59,9 @@ public class PolarisConfigBootstrapAutoConfiguration {
 	}
 
 	@Bean
-	@ConditionalOnMissingBean
-	public PolarisPropertySourceManager polarisPropertySourceManager() {
-		return new PolarisPropertySourceManager();
-	}
-
-	@Bean
 	@ConditionalOnConnectRemoteServerEnabled
 	public ConfigFileService configFileService(PolarisSDKContextManager polarisSDKContextManager) {
-		return ConfigFileServiceFactory.createConfigFileService(polarisSDKContextManager.getSDKContext());
+		return ConfigFileServiceFactory.createConfigFileService(polarisSDKContextManager.getConfigSDKContext());
 	}
 
 	@Bean
@@ -77,11 +70,9 @@ public class PolarisConfigBootstrapAutoConfiguration {
 			PolarisConfigProperties polarisConfigProperties,
 			PolarisContextProperties polarisContextProperties,
 			ConfigFileService configFileService,
-			PolarisPropertySourceManager polarisPropertySourceManager,
 			Environment environment) {
 		return new PolarisConfigFileLocator(polarisConfigProperties,
-				polarisContextProperties, configFileService,
-				polarisPropertySourceManager, environment);
+				polarisContextProperties, configFileService, environment);
 	}
 
 	@Bean