+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/endpoint/PolarisConfigEndpoint.java
@@ -37,12 +37,9 @@ import org.springframework.boot.actuate.endpoint.annotation.ReadOperation;
 public class PolarisConfigEndpoint {
 
 	private final PolarisConfigProperties polarisConfigProperties;
-	private final PolarisPropertySourceManager polarisPropertySourceManager;
 
-	public PolarisConfigEndpoint(PolarisConfigProperties polarisConfigProperties,
-			PolarisPropertySourceManager polarisPropertySourceManager) {
+	public PolarisConfigEndpoint(PolarisConfigProperties polarisConfigProperties) {
 		this.polarisConfigProperties = polarisConfigProperties;
-		this.polarisPropertySourceManager = polarisPropertySourceManager;
 	}
 
 	@ReadOperation
@@ -50,7 +47,7 @@ public class PolarisConfigEndpoint {
 		Map<String, Object> configInfo = new HashMap<>();
 		configInfo.put("PolarisConfigProperties", polarisConfigProperties);
 
-		List<PolarisPropertySource> propertySourceList = polarisPropertySourceManager.getAllPropertySources();
+		List<PolarisPropertySource> propertySourceList = PolarisPropertySourceManager.getAllPropertySources();
 		configInfo.put("PolarisPropertySource", propertySourceList);
 
 		return configInfo;