+++ spring-cloud-starter-tencent-polaris-config/src/test/java/com/tencent/cloud/polaris/config/adapter/PolarisConfigFileLocatorTest.java
@@ -29,6 +29,7 @@ import com.tencent.cloud.polaris.config.config.PolarisConfigProperties;
 import com.tencent.cloud.polaris.context.config.PolarisContextProperties;
 import com.tencent.polaris.configuration.api.core.ConfigFileService;
 import com.tencent.polaris.configuration.api.core.ConfigKVFile;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -56,14 +57,17 @@ public class PolarisConfigFileLocatorTest {
 	@Mock
 	private ConfigFileService configFileService;
 	@Mock
-	private PolarisPropertySourceManager polarisPropertySourceManager;
-	@Mock
 	private Environment environment;
+
+	@BeforeEach
+	public void setUp() {
+		PolarisPropertySourceManager.clearPropertySources();
+	}
 
 	@Test
 	public void testLoadApplicationPropertiesFile() {
 		PolarisConfigFileLocator locator = new PolarisConfigFileLocator(polarisConfigProperties, polarisContextProperties,
-				configFileService, polarisPropertySourceManager, environment);
+				configFileService, environment);
 
 		when(polarisContextProperties.getNamespace()).thenReturn(testNamespace);
 		when(polarisContextProperties.getService()).thenReturn(testServiceName);
@@ -98,7 +102,7 @@ public class PolarisConfigFileLocatorTest {
 	@Test
 	public void testActiveProfileFilesPriorityBiggerThanDefault() {
 		PolarisConfigFileLocator locator = new PolarisConfigFileLocator(polarisConfigProperties, polarisContextProperties,
-				configFileService, polarisPropertySourceManager, environment);
+				configFileService, environment);
 
 		when(polarisContextProperties.getNamespace()).thenReturn(testNamespace);
 		when(polarisContextProperties.getService()).thenReturn(testServiceName);
@@ -145,7 +149,7 @@ public class PolarisConfigFileLocatorTest {
 	@Test
 	public void testGetCustomFiles() {
 		PolarisConfigFileLocator locator = new PolarisConfigFileLocator(polarisConfigProperties, polarisContextProperties,
-				configFileService, polarisPropertySourceManager, environment);
+				configFileService, environment);
 
 		when(polarisContextProperties.getNamespace()).thenReturn(testNamespace);
 		when(polarisContextProperties.getService()).thenReturn(testServiceName);